move library to /lib
